Auto reformat & code cleanup of all custom code

+++ Scripts/Nelderim/Complete Spell System/-=+ 02 Changing/Reagents/DestroyingAngel.cs
@@ -1,45 +1,48 @@
+#region References
+
 using System;
-using Server;
+
+#endregion
 
 namespace Server.Items
 {
 	public class DestroyingAngel : BaseReagent, ICommodity
 	{
-        TextDefinition ICommodity.Description
-        {
-            get
-            {
-                return new TextDefinition( LabelNumber, String.Format( "{0} niszczejący anioł", Amount ) );
-            }
-        }
+		TextDefinition ICommodity.Description
+		{
+			get
+			{
+				return new TextDefinition(LabelNumber, String.Format("{0} niszczejący anioł", Amount));
+			}
+		}
 
-        bool ICommodity.IsDeedable { get { return false; } }
+		bool ICommodity.IsDeedable { get { return false; } }
 
 		[Constructable]
-		public DestroyingAngel() : this( 1 )
+		public DestroyingAngel() : this(1)
 		{
 		}
 
 		[Constructable]
-		public DestroyingAngel( int amount ) : base( 0xE1F, amount )
+		public DestroyingAngel(int amount) : base(0xE1F, amount)
 		{
 			Hue = 0x290;
 			Name = "niszczejący anioł";
 		}
 
-		public DestroyingAngel( Serial serial ) : base( serial )
+		public DestroyingAngel(Serial serial) : base(serial)
 		{
 		}
 
-		public override void Serialize( GenericWriter writer )
+		public override void Serialize(GenericWriter writer)
 		{
-			base.Serialize( writer );
-			writer.Write( (int) 1 ); // version
+			base.Serialize(writer);
+			writer.Write(1); // version
 		}
 
-		public override void Deserialize( GenericReader reader )
+		public override void Deserialize(GenericReader reader)
 		{
-			base.Deserialize( reader );
+			base.Deserialize(reader);
 			int version = reader.ReadInt();
 		}
 	}
