+++ Scripts/Nelderim/Complete Spell System/-=+ 02 Changing/Reagents/SpringWater.cs
@@ -1,45 +1,48 @@
+#region References
+
 using System;
-using Server;
+
+#endregion
 
 namespace Server.Items
 {
 	public class SpringWater : BaseReagent, ICommodity
 	{
-        TextDefinition ICommodity.Description
-        {
-            get
-            {
-                return new TextDefinition( LabelNumber, String.Format( "{0} wiosenna woda", Amount ) );
-            }
-        }
+		TextDefinition ICommodity.Description
+		{
+			get
+			{
+				return new TextDefinition(LabelNumber, String.Format("{0} wiosenna woda", Amount));
+			}
+		}
 
-        bool ICommodity.IsDeedable { get { return false; } }
- 
+		bool ICommodity.IsDeedable { get { return false; } }
+
 		[Constructable]
-		public SpringWater() : this( 1 )
+		public SpringWater() : this(1)
 		{
 		}
 
 		[Constructable]
-		public SpringWater( int amount ) : base( 0xE24, amount )
+		public SpringWater(int amount) : base(0xE24, amount)
 		{
 			Hue = 0x47F;
 			Name = "wiosenna woda";
 		}
 
-		public SpringWater( Serial serial ) : base( serial )
+		public SpringWater(Serial serial) : base(serial)
 		{
 		}
 
-		public override void Serialize( GenericWriter writer )
+		public override void Serialize(GenericWriter writer)
 		{
-			base.Serialize( writer );
-			writer.Write( (int) 0 ); // version
+			base.Serialize(writer);
+			writer.Write(0); // version
 		}
 
-		public override void Deserialize( GenericReader reader )
+		public override void Deserialize(GenericReader reader)
 		{
-			base.Deserialize( reader );
+			base.Deserialize(reader);
 			int version = reader.ReadInt();
 		}
 	}
