+++ Scripts/Nelderim/Complete Spell System/-=+ 03 Systems/Ancient/Scrolls/MassSleepScroll.cs
@@ -1,38 +1,36 @@
-using System;
-
 namespace Server.ACC.CSS.Systems.Ancient
 {
-    public class AncientMassSleepScroll : CSpellScroll
-    {
-        [Constructable]
-        public AncientMassSleepScroll()
-            : this(1)
-        {
-        }
+	public class AncientMassSleepScroll : CSpellScroll
+	{
+		[Constructable]
+		public AncientMassSleepScroll()
+			: this(1)
+		{
+		}
 
-        [Constructable]
-        public AncientMassSleepScroll(int amount)
-            : base(typeof(AncientMassSleepSpell), 0x1F51, amount)
-        {
-            Name = "Mass Sleep Scroll";
-            Hue = 1355;
-        }
+		[Constructable]
+		public AncientMassSleepScroll(int amount)
+			: base(typeof(AncientMassSleepSpell), 0x1F51, amount)
+		{
+			Name = "Mass Sleep Scroll";
+			Hue = 1355;
+		}
 
-        public AncientMassSleepScroll(Serial serial)
-            : base(serial)
-        {
-        }
+		public AncientMassSleepScroll(Serial serial)
+			: base(serial)
+		{
+		}
 
-        public override void Serialize(GenericWriter writer)
-        {
-            base.Serialize(writer);
-            writer.Write((int)0); // version
-        }
+		public override void Serialize(GenericWriter writer)
+		{
+			base.Serialize(writer);
+			writer.Write(0); // version
+		}
 
-        public override void Deserialize(GenericReader reader)
-        {
-            base.Deserialize(reader);
-            int version = reader.ReadInt();
-        }
-    }
+		public override void Deserialize(GenericReader reader)
+		{
+			base.Deserialize(reader);
+			int version = reader.ReadInt();
+		}
+	}
 }
