+++ Scripts/Nelderim/Complete Spell System/-=+ 03 Systems/Ancient/Scrolls/PeerScroll.cs
@@ -1,38 +1,36 @@
-using System;
-
 namespace Server.ACC.CSS.Systems.Ancient
 {
-    public class AncientPeerScroll : CSpellScroll
-    {
-        [Constructable]
-        public AncientPeerScroll()
-            : this(1)
-        {
-        }
+	public class AncientPeerScroll : CSpellScroll
+	{
+		[Constructable]
+		public AncientPeerScroll()
+			: this(1)
+		{
+		}
 
-        [Constructable]
-        public AncientPeerScroll(int amount)
-            : base(typeof(AncientPeerSpell), 0x1F43, amount)
-        {
-            Name = "Zwój wizji";
-            Hue = 1355;
-        }
+		[Constructable]
+		public AncientPeerScroll(int amount)
+			: base(typeof(AncientPeerSpell), 0x1F43, amount)
+		{
+			Name = "Zwój wizji";
+			Hue = 1355;
+		}
 
-        public AncientPeerScroll(Serial serial)
-            : base(serial)
-        {
-        }
+		public AncientPeerScroll(Serial serial)
+			: base(serial)
+		{
+		}
 
-        public override void Serialize(GenericWriter writer)
-        {
-            base.Serialize(writer);
-            writer.Write((int)0); // version
-        }
+		public override void Serialize(GenericWriter writer)
+		{
+			base.Serialize(writer);
+			writer.Write(0); // version
+		}
 
-        public override void Deserialize(GenericReader reader)
-        {
-            base.Deserialize(reader);
-            int version = reader.ReadInt();
-        }
-    }
+		public override void Deserialize(GenericReader reader)
+		{
+			base.Deserialize(reader);
+			int version = reader.ReadInt();
+		}
+	}
 }
