+++ Scripts/Nelderim/Complete Spell System/-=+ 03 Systems/Ancient/Scrolls/SleepFieldScroll.cs
@@ -1,38 +1,36 @@
-using System;
-
 namespace Server.ACC.CSS.Systems.Ancient
 {
-    public class AncientSleepFieldScroll : CSpellScroll
-    {
-        [Constructable]
-        public AncientSleepFieldScroll()
-            : this(1)
-        {
-        }
+	public class AncientSleepFieldScroll : CSpellScroll
+	{
+		[Constructable]
+		public AncientSleepFieldScroll()
+			: this(1)
+		{
+		}
 
-        [Constructable]
-        public AncientSleepFieldScroll(int amount)
-            : base(typeof(AncientSleepFieldSpell), 0x1F56, amount)
-        {
-            Name = "Sleep Field Scroll";
-            Hue = 1355;
-        }
+		[Constructable]
+		public AncientSleepFieldScroll(int amount)
+			: base(typeof(AncientSleepFieldSpell), 0x1F56, amount)
+		{
+			Name = "Sleep Field Scroll";
+			Hue = 1355;
+		}
 
-        public AncientSleepFieldScroll(Serial serial)
-            : base(serial)
-        {
-        }
+		public AncientSleepFieldScroll(Serial serial)
+			: base(serial)
+		{
+		}
 
-        public override void Serialize(GenericWriter writer)
-        {
-            base.Serialize(writer);
-            writer.Write((int)0); // version
-        }
+		public override void Serialize(GenericWriter writer)
+		{
+			base.Serialize(writer);
+			writer.Write(0); // version
+		}
 
-        public override void Deserialize(GenericReader reader)
-        {
-            base.Deserialize(reader);
-            int version = reader.ReadInt();
-        }
-    }
+		public override void Deserialize(GenericReader reader)
+		{
+			base.Deserialize(reader);
+			int version = reader.ReadInt();
+		}
+	}
 }
