+++ Scripts/Nelderim/Complete Spell System/-=+ 03 Systems/Avatar/Scrolls/AngelicFaithScroll.cs
@@ -1,36 +1,32 @@
-using System;
-using Server;
-using Server.Items;
-
 namespace Server.ACC.CSS.Systems.Avatar
 {
 	public class AvatarAngelicFaithScroll : CSpellScroll
 	{
 		[Constructable]
-		public AvatarAngelicFaithScroll() : this( 1 )
+		public AvatarAngelicFaithScroll() : this(1)
 		{
 		}
 
 		[Constructable]
-		public AvatarAngelicFaithScroll( int amount ) : base( typeof( AvatarAngelicFaithSpell ), 0xE39, amount )
+		public AvatarAngelicFaithScroll(int amount) : base(typeof(AvatarAngelicFaithSpell), 0xE39, amount)
 		{
 			Name = "Awatar Pradawnego Mnicha";
 			Hue = 1174;
 		}
 
-		public AvatarAngelicFaithScroll( Serial serial ) : base( serial )
+		public AvatarAngelicFaithScroll(Serial serial) : base(serial)
 		{
 		}
 
-		public override void Serialize( GenericWriter writer )
+		public override void Serialize(GenericWriter writer)
 		{
-			base.Serialize( writer );
-			writer.Write( (int) 0 ); // version
+			base.Serialize(writer);
+			writer.Write(0); // version
 		}
 
-		public override void Deserialize( GenericReader reader )
+		public override void Deserialize(GenericReader reader)
 		{
-			base.Deserialize( reader );
+			base.Deserialize(reader);
 			int version = reader.ReadInt();
 		}
 	}
