+++ Scripts/Nelderim/Complete Spell System/-=+ 03 Systems/Avatar/Scrolls/HeavensGateScroll.cs
@@ -1,36 +1,32 @@
-using System;
-using Server;
-using Server.Items;
-
 namespace Server.ACC.CSS.Systems.Avatar
 {
 	public class AvatarHeavensGateScroll : CSpellScroll
 	{
 		[Constructable]
-		public AvatarHeavensGateScroll() : this( 1 )
+		public AvatarHeavensGateScroll() : this(1)
 		{
 		}
 
 		[Constructable]
-		public AvatarHeavensGateScroll( int amount ) : base( typeof( AvatarHeavensGateSpell ), 0xE39, amount )
+		public AvatarHeavensGateScroll(int amount) : base(typeof(AvatarHeavensGateSpell), 0xE39, amount)
 		{
 			Name = "Niebiańska Brama";
 			Hue = 1174;
 		}
 
-		public AvatarHeavensGateScroll( Serial serial ) : base( serial )
+		public AvatarHeavensGateScroll(Serial serial) : base(serial)
 		{
 		}
 
-		public override void Serialize( GenericWriter writer )
+		public override void Serialize(GenericWriter writer)
 		{
-			base.Serialize( writer );
-			writer.Write( (int) 0 ); // version
+			base.Serialize(writer);
+			writer.Write(0); // version
 		}
 
-		public override void Deserialize( GenericReader reader )
+		public override void Deserialize(GenericReader reader)
 		{
-			base.Deserialize( reader );
+			base.Deserialize(reader);
 			int version = reader.ReadInt();
 		}
 	}
