+++ Scripts/Nelderim/Complete Spell System/-=+ 03 Systems/Bard/Scrolls/FireThrenodyScroll.cs
@@ -1,36 +1,32 @@
-using System;
-using Server;
-using Server.Items;
-
 namespace Server.ACC.CSS.Systems.Bard
 {
 	public class BardFireThrenodyScroll : CSpellScroll
 	{
 		[Constructable]
-		public BardFireThrenodyScroll() : this( 1 )
+		public BardFireThrenodyScroll() : this(1)
 		{
 		}
 
 		[Constructable]
-		public BardFireThrenodyScroll( int amount ) : base( typeof( BardFireThrenodySpell ), 0x14ED, amount )
+		public BardFireThrenodyScroll(int amount) : base(typeof(BardFireThrenodySpell), 0x14ED, amount)
 		{
 			Name = "Palący Tren";
 			Hue = 0x96;
 		}
 
-		public BardFireThrenodyScroll( Serial serial ) : base( serial )
+		public BardFireThrenodyScroll(Serial serial) : base(serial)
 		{
 		}
 
-		public override void Serialize( GenericWriter writer )
+		public override void Serialize(GenericWriter writer)
 		{
-			base.Serialize( writer );
-			writer.Write( (int) 0 ); // version
+			base.Serialize(writer);
+			writer.Write(0); // version
 		}
 
-		public override void Deserialize( GenericReader reader )
+		public override void Deserialize(GenericReader reader)
 		{
-			base.Deserialize( reader );
+			base.Deserialize(reader);
 			int version = reader.ReadInt();
 		}
 	}
