+++ Scripts/Nelderim/Complete Spell System/-=+ 03 Systems/Bard/Scrolls/PoisonCarolScroll.cs
@@ -1,36 +1,32 @@
-using System;
-using Server;
-using Server.Items;
-
 namespace Server.ACC.CSS.Systems.Bard
 {
 	public class BardPoisonCarolScroll : CSpellScroll
 	{
 		[Constructable]
-		public BardPoisonCarolScroll() : this( 1 )
+		public BardPoisonCarolScroll() : this(1)
 		{
 		}
 
 		[Constructable]
-		public BardPoisonCarolScroll( int amount ) : base( typeof( BardPoisonCarolSpell ), 0x14ED, amount )
+		public BardPoisonCarolScroll(int amount) : base(typeof(BardPoisonCarolSpell), 0x14ED, amount)
 		{
 			Name = "Wężowa Pieśń";
 			Hue = 0x96;
 		}
 
-		public BardPoisonCarolScroll( Serial serial ) : base( serial )
+		public BardPoisonCarolScroll(Serial serial) : base(serial)
 		{
 		}
 
-		public override void Serialize( GenericWriter writer )
+		public override void Serialize(GenericWriter writer)
 		{
-			base.Serialize( writer );
-			writer.Write( (int) 0 ); // version
+			base.Serialize(writer);
+			writer.Write(0); // version
 		}
 
-		public override void Deserialize( GenericReader reader )
+		public override void Deserialize(GenericReader reader)
 		{
-			base.Deserialize( reader );
+			base.Deserialize(reader);
 			int version = reader.ReadInt();
 		}
 	}
