+++ Scripts/Nelderim/Complete Spell System/-=+ 03 Systems/Cleric/Scrolls/AngelicFaithScroll.cs
@@ -1,36 +1,32 @@
-using System;
-using Server;
-using Server.Items;
-
 namespace Server.ACC.CSS.Systems.Cleric
 {
 	public class ClericAngelicFaithScroll : CSpellScroll
 	{
 		[Constructable]
-		public ClericAngelicFaithScroll() : this( 1 )
+		public ClericAngelicFaithScroll() : this(1)
 		{
 		}
 
 		[Constructable]
-		public ClericAngelicFaithScroll( int amount ) : base( typeof( ClericAngelicFaithSpell ), 0xE39, amount )
+		public ClericAngelicFaithScroll(int amount) : base(typeof(ClericAngelicFaithSpell), 0xE39, amount)
 		{
 			Name = "Anielska Wiara";
 			Hue = 0x1F0;
 		}
 
-		public ClericAngelicFaithScroll( Serial serial ) : base( serial )
+		public ClericAngelicFaithScroll(Serial serial) : base(serial)
 		{
 		}
 
-		public override void Serialize( GenericWriter writer )
+		public override void Serialize(GenericWriter writer)
 		{
-			base.Serialize( writer );
-			writer.Write( (int) 0 ); // version
+			base.Serialize(writer);
+			writer.Write(0); // version
 		}
 
-		public override void Deserialize( GenericReader reader )
+		public override void Deserialize(GenericReader reader)
 		{
-			base.Deserialize( reader );
+			base.Deserialize(reader);
 			int version = reader.ReadInt();
 		}
 	}
