+++ Scripts/Nelderim/Complete Spell System/-=+ 03 Systems/Druid/Scrolls/LureStoneScroll.cs
@@ -1,36 +1,32 @@
-using System;
-using Server;
-using Server.Items;
-
 namespace Server.ACC.CSS.Systems.Druid
 {
 	public class DruidLureStoneScroll : CSpellScroll
 	{
 		[Constructable]
-		public DruidLureStoneScroll() : this( 1 )
+		public DruidLureStoneScroll() : this(1)
 		{
 		}
 
 		[Constructable]
-		public DruidLureStoneScroll( int amount ) : base( typeof( DruidLureStoneSpell ), 0xE39, amount )
+		public DruidLureStoneScroll(int amount) : base(typeof(DruidLureStoneSpell), 0xE39, amount)
 		{
 			Name = "Ciekawy kamień";
 			Hue = 0x58B;
 		}
 
-		public DruidLureStoneScroll( Serial serial ) : base( serial )
+		public DruidLureStoneScroll(Serial serial) : base(serial)
 		{
 		}
 
-		public override void Serialize( GenericWriter writer )
+		public override void Serialize(GenericWriter writer)
 		{
-			base.Serialize( writer );
-			writer.Write( (int) 0 ); // version
+			base.Serialize(writer);
+			writer.Write(0); // version
 		}
 
-		public override void Deserialize( GenericReader reader )
+		public override void Deserialize(GenericReader reader)
 		{
-			base.Deserialize( reader );
+			base.Deserialize(reader);
 			int version = reader.ReadInt();
 		}
 	}
