+++ Scripts/Nelderim/Complete Spell System/-=+ 03 Systems/Ranger/Scrolls/FireBowScroll.cs
@@ -1,36 +1,32 @@
-using System;
-using Server;
-using Server.Items;
-
 namespace Server.ACC.CSS.Systems.Ranger
 {
 	public class RangerFireBowScroll : CSpellScroll
 	{
 		[Constructable]
-		public RangerFireBowScroll() : this( 1 )
+		public RangerFireBowScroll() : this(1)
 		{
 		}
 
 		[Constructable]
-		public RangerFireBowScroll( int amount ) : base( typeof( RangerFireBowSpell ), 3828, amount )
+		public RangerFireBowScroll(int amount) : base(typeof(RangerFireBowSpell), 3828, amount)
 		{
 			Name = "Ognisty Łuk";
 			Hue = 2001;
 		}
 
-		public RangerFireBowScroll( Serial serial ) : base( serial )
+		public RangerFireBowScroll(Serial serial) : base(serial)
 		{
 		}
 
-		public override void Serialize( GenericWriter writer )
+		public override void Serialize(GenericWriter writer)
 		{
-			base.Serialize( writer );
-			writer.Write( (int) 0 ); // version
+			base.Serialize(writer);
+			writer.Write(0); // version
 		}
 
-		public override void Deserialize( GenericReader reader )
+		public override void Deserialize(GenericReader reader)
 		{
-			base.Deserialize( reader );
+			base.Deserialize(reader);
 			int version = reader.ReadInt();
 		}
 	}
