+++ Scripts/Nelderim/Complete Spell System/-=+ 03 Systems/Undead/Scrolls/VolcanicEruptionScroll.cs
@@ -1,36 +1,32 @@
-using System;
-using Server;
-using Server.Items;
-
 namespace Server.ACC.CSS.Systems.Undead
 {
 	public class UndeadVolcanicEruptionScroll : CSpellScroll
 	{
 		[Constructable]
-		public UndeadVolcanicEruptionScroll() : this( 1 )
+		public UndeadVolcanicEruptionScroll() : this(1)
 		{
 		}
 
 		[Constructable]
-		public UndeadVolcanicEruptionScroll( int amount ) : base( typeof( UndeadVolcanicEruptionSpell ), 0xE39, amount )
+		public UndeadVolcanicEruptionScroll(int amount) : base(typeof(UndeadVolcanicEruptionSpell), 0xE39, amount)
 		{
 			Name = "Erupcja Wulkaniczna";
 			Hue = 38;
 		}
 
-		public UndeadVolcanicEruptionScroll( Serial serial ) : base( serial )
+		public UndeadVolcanicEruptionScroll(Serial serial) : base(serial)
 		{
 		}
 
-		public override void Serialize( GenericWriter writer )
+		public override void Serialize(GenericWriter writer)
 		{
-			base.Serialize( writer );
-			writer.Write( (int) 0 ); // version
+			base.Serialize(writer);
+			writer.Write(0); // version
 		}
 
-		public override void Deserialize( GenericReader reader )
+		public override void Deserialize(GenericReader reader)
 		{
-			base.Deserialize( reader );
+			base.Deserialize(reader);
 			int version = reader.ReadInt();
 		}
 	}
