+++ Scripts/Nelderim/Engines/Painting/PaintBrush.cs
@@ -1,4 +1,8 @@
+#region References
+
 using Server.Gumps;
+
+#endregion
 
 namespace Server.Items
 {
@@ -35,7 +39,7 @@ namespace Server.Items
 		{
 			base.Serialize(writer);
 
-			writer.Write((int)0); // version
+			writer.Write(0); // version
 		}
 
 		public override void Deserialize(GenericReader reader)
