+++ Scripts/Nelderim/Items/Beta/ArtifactRewards/MajorArtifactDeed.cs
@@ -2,7 +2,11 @@
 //Crafted By Broze///
 /////////////////////
 
+#region References
+
 using Server.Gumps;
+
+#endregion
 
 namespace Server.Items
 {
@@ -41,7 +45,7 @@ namespace Server.Items
 		{
 			base.Serialize(writer);
 
-			writer.Write((int)0);
+			writer.Write(0);
 		}
 
 		public override void Deserialize(GenericReader reader)
